+++ tests/unit/test_nvm_persistence.c
@@ -19,7 +19,7 @@ static uds_config_t g_cfg;
 static uint8_t g_nvm_storage[2] = {0x00, 0x00};
 
 /* Mock NVM Save */
-static int mock_nvm_save(struct uds_ctx *ctx, const uint8_t *state, uint32_t len)
+static int mock_nvm_save(struct uds_ctx *ctx, const uint8_t *state, uint16_t len)
 {
     if (len == 2) {
         g_nvm_storage[0] = state[0];
@@ -30,7 +30,7 @@ static int mock_nvm_save(struct uds_ctx *ctx, const uint8_t *state, uint32_t len
 }
 
 /* Mock NVM Load */
-static int mock_nvm_load(struct uds_ctx *ctx, uint8_t *state, uint32_t len)
+static int mock_nvm_load(struct uds_ctx *ctx, uint8_t *state, uint16_t len)
 {
     if (len == 2) {
         state[0] = g_nvm_storage[0];
